requirements.txt
python == 3.11.4
numpy == 1.24.3
pandas == 1.5.3
scipy == 1.11.3
torch == 2.1.0+cu118
scikit-learn == 1.3.0
PyWavelets == 1.4.1
astropy == 6.1
h5py == 3.7.0
geomstat == 2.5.0
